requirements.txt
python-jwt==2.0.1
python-pam==1.8.2
keyring==10.3.2
